move some stuff to authService
