+++ backend/src/auth/auth.controller.ts
@@ -13,28 +13,23 @@ import {
 	HttpStatus,
 	Query,
 	ValidationPipe,
-	NotFoundException,
 } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { UserService } from '../users/user/user.service';
 import * as bcrypt from 'bcrypt';
-import { JwtService } from '@nestjs/jwt';
 import { Response, Request } from 'express';
 import { AuthGuard } from './auth.guard';
 import { AuthService } from './auth.service';
 import { LoginUserDto } from './dto/login-user.dto';
-import { User } from '../users/user/entities/user.entity';
 import * as crypto from 'crypto';
 import { IntraTokendata } from './dto/intra-tokendata.dto';
 import { globalValidationPipeOptions } from '../main.validationpipe';
-import { QueryFailedError } from 'typeorm';
 
 @UseInterceptors(ClassSerializerInterceptor)
 @Controller()
 export class AuthController {
 	constructor(
 		private userService: UserService,
-		private jwtService: JwtService,
 		private authService: AuthService,
 		private configService: ConfigService,
 	) {}
@@ -52,76 +47,6 @@ export class AuthController {
 		);
 	}
 
-	async processUserData(
-		userData: any,
-		tokendata: IntraTokendata,
-	): Promise<string> {
-		try {
-			await this.userService.findOne({
-				where: { intra_id: userData.id },
-			});
-
-			return Promise.resolve('/?type=recurring_user');
-		} catch (e) {
-			if (e instanceof NotFoundException) {
-				const newUser = new User();
-
-				newUser.intra_id = userData.id;
-				newUser.is_intra = true;
-				newUser.name = userData.login;
-				newUser.intra_login = userData.login;
-				newUser.intra_token = tokendata.access_token;
-				newUser.intra_refresh = tokendata.refresh_token;
-				newUser.intra_expires = new Date(
-					(tokendata.created_at + tokendata.expires_in) *
-						(tokendata.created_at < 3000000000 ? 1000 : 1),
-				);
-
-				// const userWithSameName: User[] = await this.userService.findAll({
-				// 	where: { name: userData.login },
-				// });
-				// if (userWithSameName.length) {
-				// 	newUser.name += '_' + userData.id.toString();
-				// }
-
-				/*
-				// for after refactor
-				*!/
-				const newUserx = this.userService.create({
-					intra_id: userData.id,
-					is_intra: true,
-					name: userData.login,
-					intra_login: userData.login,
-					intra_token: tokendata.access_token,
-					intra_refresh: tokendata.refresh_token,
-					intra_expires:
-						(tokendata.created_at + tokendata.expires_in) *
-						(tokendata.created_at < 3000000000 ? 1000 : 1),
-				});
-				/**/
-				try {
-					await this.userService.save(newUser);
-				} catch (e) {
-					if (e instanceof QueryFailedError) {
-						newUser.name += '_' + userData.id.toString();
-						const findMatchOnId: User[] = await this.userService.findAll({
-							where: { name: newUser.name },
-						});
-						if (findMatchOnId.length) {
-							newUser.name += crypto.pseudoRandomBytes(10).toString('hex');
-						}
-						await this.userService.save(newUser);
-					} else {
-						throw new BadRequestException(
-							'something went wrong on creating the user',
-						);
-					}
-				}
-				return Promise.resolve('/?new_user');
-			}
-		}
-	}
-
 	async authenticatedThroughApi(data: IntraTokendata) {
 		const validator = new ValidationPipe(globalValidationPipeOptions());
 		try {
@@ -135,7 +60,9 @@ export class AuthController {
 				},
 			})
 				.then((response) => response.json())
-				.then(async (userData) => this.processUserData(userData, stripped));
+				.then(async (userData) =>
+					this.authService.processUserData(userData, stripped),
+				);
 		} catch (e) {
 			throw new BadRequestException(
 				'Bad data recieved from intra api in authentication flow',
@@ -197,8 +124,7 @@ export class AuthController {
 			) {
 				throw new BadRequestException('Invalid user/password combination');
 			}
-			const jwt = await this.jwtService.signAsync({ id: user.id });
-			response.cookie('jwt', jwt, { httpOnly: true });
+			await this.authService.login(user.id, response);
 			return user;
 		} catch (e) {
 			throw new BadRequestException('Invalid user/password combination');
